requirements.txt
pynput
pyautogui
yagmail
python-dotenv